// apps/sales-copilot/src/generated/services/Crf5c_contactsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Crf5c_contactsBase, Crf5c_contacts } from '../models/Crf5c_contactsModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Crf5c_contactsService {
  private static readonly dataSourceName = 'crf5c_contacts';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Crf5c_contactsBase, 'crf5c_contactid'>): Promise<IOperationResult<Crf5c_contacts>> {
    const result = await Crf5c_contactsService.client.createRecordAsync<Omit<Crf5c_contactsBase, 'crf5c_contactid'>, Crf5c_contacts>(
      Crf5c_contactsService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Crf5c_contactsBase, 'crf5c_contactid'>>): Promise<IOperationResult<Crf5c_contacts>> {
    const result = await Crf5c_contactsService.client.updateRecordAsync<Partial<Omit<Crf5c_contactsBase, 'crf5c_contactid'>>, Crf5c_contacts>(
      Crf5c_contactsService.dataSourceName,
      id.toString(),
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Crf5c_contactsService.client.deleteRecordAsync(
      Crf5c_contactsService.dataSourceName,
      id.toString());
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Crf5c_contacts>> {
    const result = await Crf5c_contactsService.client.retrieveRecordAsync<Crf5c_contacts>(
      Crf5c_contactsService.dataSourceName,
      id.toString(),
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Crf5c_contacts[]>> {
    const result = await Crf5c_contactsService.client.retrieveMultipleRecordsAsync<Crf5c_contacts>(
      Crf5c_contactsService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Crf5c_contacts> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Crf5c_contactsService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Crf5c_contactsService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}