// apps/sales-copilot/src/generated/models/Crf5c_contactsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf5c_contactsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf5c_contactsstatecode = keyof typeof Crf5c_contactsstatecode;
export const Crf5c_contactsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf5c_contactsstatuscode = keyof typeof Crf5c_contactsstatuscode;

export interface Crf5c_contactsBase {
  "crf5c_Account@odata.bind": string;
  crf5c_contactid: string;
  crf5c_email?: string;
  crf5c_fullname: string;
  crf5c_phone?: string;
  crf5c_title?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf5c_contactsstatecode;
  statuscode?: Crf5c_contactsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Crf5c_contacts extends Crf5c_contactsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf5c_accountname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  crf5c_account?: object;
  _crf5c_account_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}